Restructure files to make a package
The files are moved and modified to make a package with distutils.
I tested that the package can be built and installed using SWIG 4.0.0. Derivative files, pyheclib.py and _pyheclib.*.pyd that Nicky created are kept in the repository for convenience though those can be generated by SWIG.
The package can be installed by: 'python setup.py install', assuming that the usre has a compiler and SWIG installed on the system for distutils. After the package is installed, the package can be used like any others by just 'import pyhecdss'.
Some files are linted as well.